Move model_definition
